Updating 10.0.8

+++ src/Entity.php
@@ -7,7 +7,7 @@
  *
  * http://glpi-project.org
  *
- * @copyright 2015-2022 Teclib' and contributors.
+ * @copyright 2015-2023 Teclib' and contributors.
  * @copyright 2003-2014 by the INDEPNET Development Team.
  * @licence   https://www.gnu.org/licenses/gpl-3.0.html
  *
@@ -67,9 +67,12 @@ class Entity extends CommonTreeDropdown
     /**
      * Possible values for "anonymize_support_agents" setting
      */
-    const ANONYMIZE_DISABLED     = 0;
-    const ANONYMIZE_USE_GENERIC  = 1;
-    const ANONYMIZE_USE_NICKNAME = 2;
+    const ANONYMIZE_DISABLED            = 0;
+    const ANONYMIZE_USE_GENERIC         = 1;
+    const ANONYMIZE_USE_NICKNAME        = 2;
+    const ANONYMIZE_USE_GENERIC_USER    = 3;
+    const ANONYMIZE_USE_NICKNAME_USER   = 4;
+    const ANONYMIZE_USE_GENERIC_GROUP   = 5;
 
    // Array of "right required to update" => array of fields allowed
    // Missing field here couldn't be update (no right)
@@ -152,19 +155,55 @@ class Entity extends CommonTreeDropdown
         ];
     }
 
-    /**
-     * @since 0.84
-     **/
+    public function pre_updateInDB()
+    {
+        global $DB;
+
+        if (($key = array_search('name', $this->updates)) !== false) {
+            /// Check if entity does not exist
+            $iterator = $DB->request([
+                'FROM' => $this->getTable(),
+                'WHERE' => [
+                    'name' => $this->input['name'],
+                    'entities_id' => $this->input['entities_id'],
+                    'id' => ['<>', $this->input['id']]
+                ]
+            ]);
+
+            if (count($iterator)) {
+                //To display a message
+                $this->fields['name'] = $this->oldvalues['name'];
+                unset($this->updates[$key]);
+                unset($this->oldvalues['name']);
+                Session::addMessageAfterRedirect(
+                    __('An entity with that name already exists at the same level.'),
+                    false,
+                    ERROR
+                );
+            }
+        }
+    }
+
     public function pre_deleteItem()
     {
         global $GLPI_CACHE;
 
-       // Security do not delete root entity
+        // Security do not delete root entity
         if ($this->input['id'] == 0) {
             return false;
         }
 
-       //Cleaning sons calls getAncestorsOf and thus... Re-create cache. Call it before clean.
+        // Security do not delete entity with children
+        if (countElementsInTable($this->getTable(), ['entities_id' => $this->input['id']])) {
+            Session::addMessageAfterRedirect(
+                __('You cannot delete an entity which contains sub-entities.'),
+                false,
+                ERROR
+            );
+            return false;
+        }
+
+        //Cleaning sons calls getAncestorsOf and thus... Re-create cache. Call it before clean.
         $this->cleanParentsSons();
         $ckey = 'ancestors_cache_' . $this->getTable() . '_' . $this->getID();
         $GLPI_CACHE->delete($ckey);
@@ -178,6 +217,11 @@ class Entity extends CommonTreeDropdown
         return _n('Entity', 'Entities', $nb);
     }
 
+    public static function canCreate()
+    {
+        // Do not show the create button if no recusive access on current entity
+        return parent::canCreate() && Session::haveRecursiveAccessToEntity(Session::getActiveEntity());
+    }
 
     public function canCreateItem()
     {
@@ -545,24 +589,13 @@ class Entity extends CommonTreeDropdown
 
     public function post_getFromDB()
     {
-        // Copy config "strategy" fields in corresponding "id" field.
-        if (($this->fields['calendars_strategy'] ?? 0) < 0) {
-            $this->fields['calendars_id'] = $this->fields['calendars_strategy'];
-        }
-        if (($this->fields['changetemplates_strategy'] ?? 0) < 0) {
-            $this->fields['changetemplates_id'] = $this->fields['changetemplates_strategy'];
-        }
-        if (($this->fields['contracts_strategy_default'] ?? 0) < 0) {
-            $this->fields['contracts_id_default'] = $this->fields['contracts_strategy_default'];
-        }
-        if (($this->fields['entities_strategy_software'] ?? 0) < 0) {
-            $this->fields['entities_id_software'] = $this->fields['entities_strategy_software'];
-        }
-        if (($this->fields['problemtemplates_strategy'] ?? 0) < 0) {
-            $this->fields['problemtemplates_id'] = $this->fields['problemtemplates_strategy'];
-        }
-        if (($this->fields['tickettemplates_strategy'] ?? 0) < 0) {
-            $this->fields['tickettemplates_id'] = $this->fields['tickettemplates_strategy'];
+        // Copy config "strategy" fields in corresponding "id" field
+        // when "strategy" is < 0.
+        foreach ($this->fields as $field_key => $value) {
+            if (preg_match('/_strategy(_.+)?/', $field_key) === 1 && $value < 0) {
+                $id_field_key = str_replace('_strategy', '_id', $field_key);
+                $this->fields[$id_field_key] = $this->fields[$field_key];
+            }
         }
     }
 
@@ -574,6 +607,8 @@ class Entity extends CommonTreeDropdown
        // Add right to current user - Hack to avoid login/logout
         $_SESSION['glpiactiveentities'][$this->fields['id']] = $this->fields['id'];
         $_SESSION['glpiactiveentities_string']              .= ",'" . $this->fields['id'] . "'";
+        // Root entity cannot be deleted, so if we added an entity this means GLPI is now multi-entity
+        $_SESSION['glpi_multientitiesmode'] = 1;
 
        // clean entity tree cache
         $this->cleanEntitySelectorCache();
@@ -1246,6 +1281,20 @@ class Entity extends CommonTreeDropdown
         ];
 
         $tab[] = [
+            'id'                 => '75',
+            'table'              => self::getTable(),
+            'field'              => 'contracts_id_default',
+            'name'               => __('Default contract'),
+            'datatype'           => 'specific',
+            'nosearch'           => true,
+            'additionalfields'   => ['contracts_strategy_default'],
+            'toadd'              => [
+                self::CONFIG_PARENT => __('Inheritance of the parent entity'),
+                self::CONFIG_AUTO   => __('Contract in ticket entity'),
+            ]
+        ];
+
+        $tab[] = [
             'id'                 => 'assets',
             'name'               => _n('Asset', 'Assets', Session::getPluralNumber())
         ];
@@ -1885,12 +1934,22 @@ class Entity extends CommonTreeDropdown
         $params = [
             'name'       => 'transfers_id',
             'value'      => $entity->fields['transfers_id'],
-            'emptylabel' => __('No automatic transfer')
+            'display_emptychoice' => false
         ];
         if ($entity->fields['id'] > 0) {
-            $params['toadd'] = [self::CONFIG_PARENT => __('Inheritance of the parent entity')];
+            $params['toadd'] = [
+                self::CONFIG_NEVER => __('No automatic transfer'),
+                self::CONFIG_PARENT => __('Inheritance of the parent entity')
+            ];
         }
         Dropdown::show('Transfer', $params);
+        if ($entity->fields['transfers_strategy'] == self::CONFIG_PARENT) {
+            $inherited_strategy = self::getUsedConfig('transfers_strategy', $entity->fields['entities_id']);
+            $inherited_value    = $inherited_strategy === 0
+                ? self::getUsedConfig('transfers_strategy', $entity->fields['entities_id'], 'transfers_id')
+                : $inherited_strategy;
+            self::inheritedValue(self::getSpecificValueToDisplay('transfers_id', $inherited_value));
+        }
         echo "</td>";
         echo "</td><td colspan='2'></td></tr>";
 
@@ -1954,9 +2013,12 @@ class Entity extends CommonTreeDropdown
         echo "<tr class='tab_bg_1'>";
         echo "<td>" . __('Administrator email address') . "</td>";
         echo "<td>";
-        echo Html::input('admin_email', ['value' => $entity->fields['admin_email']]);
+        echo Html::input('admin_email', ['value' => $entity->fields['admin_email'], 'type' => 'email']);
         if (empty($entity->fields['admin_email']) && $ID > 0) {
             self::inheritedValue(self::getUsedConfig('admin_email', $ID, '', ''));
+        }
+        if (!empty($entity->fields['admin_email']) && !NotificationMailing::isUserAddressValid($entity->fields['admin_email'])) {
+            echo "<span class='red'>" . __('Invalid email address') . "</span>";
         }
         echo "</td>";
         echo "<td>" . __('Administrator name') . "</td><td>";
@@ -1971,9 +2033,12 @@ class Entity extends CommonTreeDropdown
         echo "<tr class='tab_bg_1'>";
         echo "<td>" . __('Email sender address') . "</td>";
         echo "<td>";
-        echo Html::input('from_email', ['value' => $entity->fields['from_email']]);
+        echo Html::input('from_email', ['value' => $entity->fields['from_email'], 'type' => 'email']);
         if (empty($entity->fields['from_email']) && $ID > 0) {
             self::inheritedValue(self::getUsedConfig('from_email', $ID, '', ''));
+        }
+        if (!empty($entity->fields['from_email']) && !NotificationMailing::isUserAddressValid($entity->fields['from_email'])) {
+            echo "<span class='red'>" . __('Invalid email address') . "</span>";
         }
         echo "</td>";
         echo "<td>" . __('Email sender name') . "</td><td>";
@@ -1988,9 +2053,12 @@ class Entity extends CommonTreeDropdown
         echo "<tr class='tab_bg_1'>";
         echo "<td>" . __('No-Reply address') . "</td>";
         echo "<td>";
-        echo Html::input('noreply_email', ['value' => $entity->fields['noreply_email']]);
+        echo Html::input('noreply_email', ['value' => $entity->fields['noreply_email'], 'type' => 'email']);
         if (empty($entity->fields['noreply_email']) && $ID > 0) {
             self::inheritedValue(self::getUsedConfig('noreply_email', $ID, '', ''));
+        }
+        if (!empty($entity->fields['noreply_email']) && !NotificationMailing::isUserAddressValid($entity->fields['noreply_email'])) {
+            echo "<span class='red'>" . __('Invalid email address') . "</span>";
         }
         echo "</td>";
         echo "<td>" . __('No-Reply name') . "</td><td>";
@@ -2005,9 +2073,12 @@ class Entity extends CommonTreeDropdown
         echo "<tr class='tab_bg_1'>";
         echo "<td><label for='replyto_email'>" . __('Reply-To address') . "</label></td>";
         echo "<td>";
-        echo Html::input('replyto_email', ['value' => $entity->fields['replyto_email']]);
+        echo Html::input('replyto_email', ['value' => $entity->fields['replyto_email'], 'type' => 'email']);
         if (empty($entity->fields['replyto_email']) && $ID > 0) {
             self::inheritedValue(self::getUsedConfig('replyto_email', $ID, '', ''));
+        }
+        if (!empty($entity->fields['replyto_email']) && !NotificationMailing::isUserAddressValid($entity->fields['replyto_email'])) {
+            echo "<span class='red'>" . __('Invalid email address') . "</span>";
         }
         echo "</td>";
         echo "<td><label for='replyto_email_name'>" . __('Reply-To name') . "</label></td>";
@@ -2929,7 +3000,7 @@ class Entity extends CommonTreeDropdown
 
         Contract::dropdown([
             'name'      => 'contracts_id_default',
-            'condition' => Contract::getExpiredCriteria(),
+            'condition' => ['is_template' => 0, 'is_deleted' => 0] + Contract::getExpiredCriteria(),
             'entity'    => $entity->getID(),
             'toadd'     => $toadd,
             'value'     => $current_default_contract_value,
@@ -3478,8 +3549,11 @@ class Entity extends CommonTreeDropdown
         return [
             self::CONFIG_PARENT => __('Inheritance of the parent entity'),
             self::ANONYMIZE_DISABLED => __('Disabled'),
-            self::ANONYMIZE_USE_GENERIC => __("Replace the agent's name with a generic name"),
-            self::ANONYMIZE_USE_NICKNAME => __("Replace the agent's name with a customisable nickname"),
+            self::ANONYMIZE_USE_GENERIC => __("Replace the agent and group name with a generic name"),
+            self::ANONYMIZE_USE_NICKNAME => __("Replace the agent and group name with a customisable nickname"),
+            self::ANONYMIZE_USE_GENERIC_USER => __("Replace the agent's name with a generic name"),
+            self::ANONYMIZE_USE_NICKNAME_USER => __("Replace the agent's name with a customisable nickname"),
+            self::ANONYMIZE_USE_GENERIC_GROUP => __("Replace the group's name with a generic name"),
         ];
     }
 
@@ -3684,6 +3758,28 @@ class Entity extends CommonTreeDropdown
                     return __('24/7');
                 }
                 return Dropdown::getDropdownName('glpi_calendars', $values[$field]);
+
+            case 'transfers_id':
+                $strategy = $values['transfers_strategy'] ?? $values[$field];
+                if ($strategy == self::CONFIG_NEVER) {
+                }
+                if ($strategy == self::CONFIG_PARENT) {
+                    return __('Inheritance of the parent entity');
+                } elseif ($values[$field] == 0) {
+                    return __('No automatic transfer');
+                }
+                return Dropdown::getDropdownName('glpi_transfers', $values[$field]);
+
+            case 'contracts_id_default':
+                $strategy = $values['contracts_strategy_default'] ?? $values[$field];
+                if ($strategy === self::CONFIG_PARENT) {
+                    return __('Inheritance of the parent entity');
+                }
+                if ($strategy === self::CONFIG_AUTO) {
+                    return __('Contract in ticket entity');
+                }
+
+                return Dropdown::getDropdownName(Contract::getTable(), $values[$field]);
         }
         return parent::getSpecificValueToDisplay($field, $values, $options);
     }
@@ -3866,6 +3962,8 @@ class Entity extends CommonTreeDropdown
      * @since 10.0.0
      *
      * @return array
+     *
+     * @FIXME Remove this method in GLPI 10.1.
      */
     public static function getDefaultContractValues($entities_id): array
     {
@@ -3876,6 +3974,8 @@ class Entity extends CommonTreeDropdown
 
         $contract = new Contract();
         $criteria = getEntitiesRestrictCriteria('', '', $entities_id, true);
+        $criteria['is_deleted'] = 0;
+        $criteria['is_template'] = 0;
         $criteria[] = Contract::getExpiredCriteria();
         $contracts = $contract->find($criteria);
 
@@ -3901,7 +4001,11 @@ class Entity extends CommonTreeDropdown
         if ($entity_default_contract_strategy == self::CONFIG_AUTO) {
             // Contract in current entity
             $contract = new Contract();
-            $criteria = ['entities_id' => $entities_id];
+            $criteria = [
+                'entities_id' => $entities_id,
+                'is_deleted'  => 0,
+                'is_template' => 0,
+            ];
             $criteria[] = Contract::getExpiredCriteria();
             $contracts = $contract->find($criteria);
 
@@ -3917,7 +4021,14 @@ class Entity extends CommonTreeDropdown
         return self::getUsedConfig('contracts_strategy_default', $entities_id, 'contracts_id_default', 0);
     }
 
-    public static function badgeCompletename(string $entity_string = ""): string
+    /**
+     * Return HTML code for entity badge showing its completename.
+     *
+     * @param string $entity_string
+     *
+     * @return string|null
+     */
+    public static function badgeCompletename(string $entity_string = "", ?string $title = null): string
     {
         // `completename` is expected to be received as it is stored in DB,
         // meaning that `>` separator is not encoded, but `<`, `>` and `&` from self or parent names are encoded.
@@ -3929,7 +4040,9 @@ class Entity extends CommonTreeDropdown
         }
 
         // Construct HTML with special chars encoded.
-        $title = htmlspecialchars(implode(' > ', $names));
+        if ($title === null) {
+            $title = htmlspecialchars(implode(' > ', $names));
+        }
         $breadcrumbs = implode(
             '<i class="fas fa-caret-right mx-1"></i>',
             array_map(
@@ -3941,7 +4054,7 @@ class Entity extends CommonTreeDropdown
         );
 
 
-        return '<span class="entity-badge" title="' . $title . '">' . $breadcrumbs . "</span>";
+        return '<span class="glpi-badge" title="' . $title . '">' . $breadcrumbs . "</span>";
     }
 
     /**
@@ -3956,6 +4069,57 @@ class Entity extends CommonTreeDropdown
         $entity = new self();
         if ($entity->getFromDB($entity_id)) {
             return self::badgeCompletename($entity->fields['completename']);
+        }
+        return null;
+    }
+
+    /**
+     * Return HTML code for entity badge showing its completename with last entity as HTML link.
+     *
+     * @param object $entity
+     *
+     * @return string|null
+     */
+    public static function badgeCompletenameLink(object $entity): string
+    {
+        // `completename` is expected to be received as it is stored in DB,
+        // meaning that `>` separator is not encoded, but `<`, `>` and `&` from self or parent names are encoded.
+        $names = explode(' > ', trim($entity->fields['completename']));
+        // Convert the whole completename into decoded HTML.
+        foreach ($names as &$name) {
+            $name = Sanitizer::decodeHtmlSpecialChars($name);
+        }
+
+        // Construct HTML with special chars encoded.
+        $title       = htmlspecialchars(implode(' > ', $names));
+        $last_name   = array_pop($names);
+        $breadcrumbs = implode(
+            '<i class="fas fa-caret-right mx-1"></i>',
+            array_map(
+                function (string $name): string {
+                    return '<span class="text-nowrap text-muted">' . htmlspecialchars($name) . '</span>';
+                },
+                $names
+            )
+        );
+
+        $last_url  = '<i class="fas fa-caret-right mx-1"></i>' . '<a href="' . $entity->getLinkURL() . '" title="' . $title . '">' . htmlspecialchars($last_name) . '</a>';
+
+        return '<span class="glpi-badge" title="' . $title . '">' . $breadcrumbs . $last_url . '</span>';
+    }
+
+    /**
+     * Return HTML code for entity badge showing its completename with last entity as HTML link.
+     *
+     * @param int $entity_id
+     *
+     * @return string|null
+     */
+    public static function badgeCompletenameLinkById(int $entity_id): ?string
+    {
+        $entity = new self();
+        if ($entity->getFromDB($entity_id)) {
+            return self::badgeCompletenameLink($entity);
         }
         return null;
     }
